Rare: Move the different applications into their respective folders
