+++ rare/launcher/lgd_helper.py
